+++ src/main/java/roomescape/controller/api/AuthController.java
@@ -7,9 +7,9 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.controller.exception.UnauthorizedException;
 import roomescape.dto.request.LoginRequest;
 import roomescape.dto.response.MemberResponse;
+import roomescape.security.exception.UnauthorizedException;
 import roomescape.service.AuthService;
 import roomescape.service.MemberService;
 import roomescape.util.CookieUtil;
