+++ src/main/java/roomescape/controller/api/ReservationController.java
@@ -16,11 +16,11 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.controller.dto.response.ApiResponses;
+import roomescape.controller.support.Auth;
 import roomescape.dto.request.ReservationRequest;
 import roomescape.dto.response.PersonalReservationResponse;
 import roomescape.dto.response.ReservationResponse;
 import roomescape.security.Accessor;
-import roomescape.security.Auth;
 import roomescape.service.ReservationService;
 
 @RestController
