+++ src/main/java/roomescape/controller/api/ReservationWaitingController.java
@@ -13,10 +13,10 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.controller.support.Auth;
 import roomescape.dto.request.ReservationRequest;
 import roomescape.dto.response.ReservationResponse;
 import roomescape.security.Accessor;
-import roomescape.security.Auth;
 import roomescape.service.ReservationWaitingService;
 
 @RestController
