+++ src/main/java/roomescape/controller/support/Auth.java
@@ -1,4 +1,4 @@
-package roomescape.security;
+package roomescape.controller.support;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
