+++ src/main/java/roomescape/controller/support/AuthArgumentResolver.java
@@ -1,4 +1,4 @@
-package roomescape.security;
+package roomescape.controller.support;
 
 import org.springframework.core.MethodParameter;
 import org.springframework.stereotype.Component;
@@ -6,8 +6,10 @@ import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
-import roomescape.controller.exception.UnauthorizedException;
 import roomescape.dto.response.MemberResponse;
+import roomescape.security.Accessor;
+import roomescape.security.TokenProvider;
+import roomescape.security.exception.UnauthorizedException;
 import roomescape.service.MemberService;
 import roomescape.util.CookieUtil;
