+++ src/main/java/roomescape/infrastructure/JwtTokenProvider.java
@@ -1,4 +1,4 @@
-package roomescape.security;
+package roomescape.infrastructure;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.ExpiredJwtException;
@@ -12,6 +12,7 @@ import java.util.Date;
 import javax.crypto.SecretKey;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
+import roomescape.security.TokenProvider;
 
 @Component
 public class JwtTokenProvider implements TokenProvider {
