+++ src/main/java/roomescape/security/exception/UnauthorizedException.java
@@ -1,4 +1,4 @@
-package roomescape.controller.exception;
+package roomescape.security.exception;
 
 public class UnauthorizedException extends RuntimeException {
