+++ src/test/java/roomescape/infrastructure/JwtTokenProviderTest.java
@@ -1,4 +1,4 @@
-package roomescape.security;
+package roomescape.infrastructure;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
